#2326 Fix  Meanings: Linked variable templates do not block delete. Do not return 500, but error message.

+++ Console/BExIS.Web.Shell/Areas/RPM/BExIS.Modules.Rpm.UI/Controllers/MeaningController.cs
@@ -94,6 +94,9 @@ namespace BExIS.Modules.Rpm.UI.Controllers
             {
                 using (var _meaningManager = new MeaningManager())
                 {
+                    if (_meaningManager.IsMeaningInUse(id))
+                        return Json(false, JsonRequestBehavior.AllowGet);
+
                     _meaningManager.DeleteMeaning(id);
                     return Json(true, JsonRequestBehavior.AllowGet);
                 }
